+++ app/grails-app/domain/com/k_int/kbplus/Package.groovy
@@ -444,17 +444,17 @@ static hasMany = [  tipps:     TitleInstancePackagePlatform,
                      platformId:tip.platform.id,
                      platformUuid:tip.platform.gokbId,
                      coverage:[],
-                     url:tip.hostPlatformURL,
+                     url:tip.hostPlatformURL ?: '',
                      identifiers:[],
                      status: tip.status,
-                     accessStart: tip.accessStartDate,
-                     accessEnd: tip.accessEndDate
+                     accessStart: tip.accessStartDate ?: null,
+                     accessEnd: tip.accessEndDate ?: null
                    ];
 
       // Need to format these dates using correct mask
       newtip.coverage.add([
-                        startDate:tip.startDate ? sdf.format(tip.startDate) : '',
-                        endDate:tip.endDate ? sdf.format(tip.endDate) : '',
+                        startDate:tip.startDate ?: null,
+                        endDate:tip.endDate ?: null,
                         startVolume:tip.startVolume ?: '',
                         endVolume:tip.endVolume ?: '',
                         startIssue:tip.startIssue ?: '',