+++ app/grails-app/services/de/laser/StatsSyncService.groovy
@@ -1,9 +1,11 @@
 package de.laser
 
 import com.k_int.kbplus.*
-import com.k_int.properties.*
 import de.laser.domain.StatsTripleCursor
+import de.laser.usage.StatsSyncServiceOptions
+import de.laser.usage.SushiClient
 import groovy.json.JsonOutput
+import groovy.time.TimeCategory
 import groovyx.net.http.*
 
 import java.security.MessageDigest
@@ -11,13 +13,12 @@ import java.text.SimpleDateFormat
 import java.time.YearMonth
 import java.time.format.DateTimeFormatter
 
-import static groovyx.net.http.ContentType.*
 import groovyx.gpars.GParsPool
 
 class StatsSyncService {
 
     static final THREAD_POOL_SIZE = 4
-    static final SYNC_STATS_FROM = '2012-01'
+    static final SYNC_STATS_FROM = '2012-01-01'
 
     def grailsApplication
     def executorService
@@ -63,8 +64,7 @@ class StatsSyncService {
         availableReportCache = [:]
     }
 
-    private String getTitleInstancesForUsageQuery()
-    {
+    private String getTitleInstancesForUsageQuery() {
         // Distinct list of titles ids, the content provider, subscribing organisation and the zdbid
        def hql =  "select distinct ie.tipp.title.id, po.org.id, orgrel.org.id, zdbtitle.id from IssueEntitlement as ie " +
             "join ie.tipp.pkg.orgs as po " +
@@ -83,8 +83,7 @@ class StatsSyncService {
         return hql
     }
 
-    def addFilters(params)
-    {
+    def addFilters(params) {
         queryParams = [:]
         if (params.supplier != 'null'){
             queryParams['supplier'] = params.supplier as long
@@ -92,25 +91,6 @@ class StatsSyncService {
         if (params.institution != 'null'){
             queryParams['institution'] = params.institution as long
         }
-    }
-
-    private getReportType(report) {
-        def result
-        switch (report) {
-            case "JR1":
-                result = "journal"
-            break
-            case "JR1GOA":
-                result = "journal"
-            break
-            case "DB1":
-                result = "database"
-            break
-            default:
-                result = "journal"
-            break
-        }
-        return result
     }
 
     def doSync() {
@@ -205,146 +185,86 @@ class StatsSyncService {
         }
     }
 
+    ArrayList getObjectsForItem(listItem) {
+       [
+            TitleInstance.get(listItem[0]),
+            Org.get(listItem[1]),
+            Org.get(listItem[2]),
+            IdentifierOccurrence.get(listItem[3])
+        ]
+    }
+
+    def getRelevantReportList(queryParams) {
+        def reports = RefdataValue.findAllByOwner(RefdataCategory.findByDesc('FactType'))
+        def availableReports = getAvailableReportsForPlatform(queryParams)
+        reports.removeAll {
+            if (it.value.startsWith('STATS') || (it.value.startsWith('JUSP'))){
+                log.warn('STATS/JUSP prefix deprecated please remove Refdatavalues')
+            }
+            def reportInAvailableReport = it.value in availableReports
+            (it.value.startsWith('STATS') || it.value.startsWith('JUSP') || !reportInAvailableReport)
+        }
+        return reports
+    }
+
+    StatsSyncServiceOptions initializeStatsSyncServiceOptions(listItem, mostRecentClosedPeriod) {
+        def options = new StatsSyncServiceOptions()
+        def itemObjects = getObjectsForItem(listItem)
+        options.setItemObjects(itemObjects)
+        options.setBasicQueryParams()
+        options.mostRecentClosedPeriod = mostRecentClosedPeriod
+        return options
+    }
+
+    def getCursor(options) {
+        // There could be more than one (if we have gaps in usage), get the newest one
+        def csr = StatsTripleCursor.findByTitleIdAndSupplierIdAndCustomerIdAndFactType(
+            options.statsTitleIdentifier, options.platform, options.customer, options.factType,
+            [sort: "availTo", order: "desc"])
+        if (csr == null) {
+            csr = new StatsTripleCursor(
+                titleId: options.statsTitleIdentifier,
+                supplierId: options.platform,
+                customerId: options.customer,
+                availFrom: new SimpleDateFormat("yyyy-MM-dd").parse(SYNC_STATS_FROM),
+                availTo: null,
+                factType: options.factType
+            )
+            csr.numFacts = 0
+        }
+        return csr
+    }
+
     def processListItem(listItem, mostRecentClosedPeriod) {
-        def uri = new URIBuilder(grailsApplication.config.statsApiUrl)
-        def baseUrl = uri.getScheme()+"://"+uri.getHost()
-        def basePath = uri.getPath().endsWith('/') ? uri.getPath() : uri.getPath() + '/'
-        def path = basePath + 'Sushiservice/GetReport'
-        def stats_api_endpoint = new RESTClient(baseUrl)
-        def timeStampFormat = new SimpleDateFormat('yyyy-MM-dd')
+        def sushiClient = new SushiClient()
         def start_time = System.currentTimeMillis()
 
         Fact.withNewTransaction { status ->
-
-            def title_inst = TitleInstance.get(listItem[0])
-            def supplier_inst = Org.get(listItem[1])
-            def org_inst = Org.get(listItem[2])
-            def title_io_inst = IdentifierOccurrence.get(listItem[3])
-            def statsTitleIdentifier = title_io_inst.identifier.value
-
-            def platform = supplier_inst.getIdentifierByType('statssid').value
-            def customer = org_inst.getIdentifierByType('wibid').value
-            def apiKey = OrgCustomProperty.findByTypeAndOwner(PropertyDefinition.findByName("API Key"), org_inst)
-            def requestor = OrgCustomProperty.findByTypeAndOwner(PropertyDefinition.findByName("RequestorID"), org_inst)
-            def queryParams = [platform:platform, customer:customer, apiKey: apiKey, requestor:requestor]
-            def availableReports = getAvailableReportsForPlatform(queryParams)
-            def reports = RefdataValue.findAllByOwner(RefdataCategory.findByDesc('FactType'))
-            reports.removeAll {
-                if (it.value.startsWith('STATS')){
-                    log.warn('STATS prefix deprecated please remove Refdatavalues')
-                }
-                def reportInAvailableReport = it.value in availableReports
-                (it.value.startsWith('STATS') || !reportInAvailableReport)
-            }
-
-            def csr = null
+            def options = initializeStatsSyncServiceOptions(listItem, mostRecentClosedPeriod)
+            def reports = getRelevantReportList(options.getBasicQueryParams())
+            StatsTripleCursor csr = null
 
             reports.each { statsReport ->
-                def factCount = 0
-                def matcher = statsReport.value =~ /^(.*).(\d)$/
-                def report = matcher[0][1]
-                def version = matcher[0][2]
-                def reportType = getReportType(report)
-                def titleId = title_io_inst.identifier.value
-                def factType = RefdataCategory.lookupOrCreate('FactType', statsReport.toString())
-
+                options.setReportSpecificQueryParams(statsReport)
                 // we could use a more complex structure, e.g. to try to seperate the SUSHI Exceptions from API
                 // for now use a list of error messages
                 def jsonErrors = []
-                csr = StatsTripleCursor.findByTitleIdAndSupplierIdAndCustomerIdAndFactType(statsTitleIdentifier, platform, customer, factType)
-                if (csr == null) {
-                    csr = new StatsTripleCursor(
-                        titleId: statsTitleIdentifier,
-                        supplierId: platform,
-                        customerId: customer,
-                        haveUpTo: null,
-                        factType: factType
-                    )
-                    csr.numFacts = 0
-                }
-                if ((csr.haveUpTo == null) || (csr.haveUpTo < mostRecentClosedPeriod)) {
-                    def fromPeriod = SYNC_STATS_FROM
-                    if (csr.haveUpTo){
-                        DateTimeFormatter formatter = DateTimeFormatter.ofPattern('yyyy-MM')
-                        YearMonth localDate = YearMonth.parse(csr.haveUpTo, formatter)
-                        fromPeriod = localDate.plusMonths(1).toString()
-                    }
+                csr = getCursor(options)
+                def mostRecentClosedDate = new SimpleDateFormat("yyyy-MM-dd").parse(options.mostRecentClosedPeriod)
+                if ((csr.availTo == null) || (csr.availTo < mostRecentClosedDate)) {
+                    options.from = getNextFromPeriod(csr)
+                    sushiClient.clientOptions = options
                     try {
-                        def beginDate = "${fromPeriod}-01"
-                        def endDate = getLastDayOfMonth(mostRecentClosedPeriod)
-                        log.debug("Calling STATS API:  ${report}, Title with ID ${titleId}")
-                        log.debug("Period Begin: ${beginDate}, Period End: ${endDate}")
-                        stats_api_endpoint.get(
-                                path: path,
-                                contentType: ANY, // We get no XmlSlurper Objects for value XML
-                                query: [
-                                        APIKey        : apiKey,
-                                        RequestorID   : requestor,
-                                        CustomerID    : customer,
-                                        Report        : report,
-                                        Release       : version,
-                                        BeginDate     : beginDate,
-                                        EndDate       : endDate,
-                                        Platform      : platform,
-                                        ItemIdentifier: "${reportType}:zdbid:" + titleId
-                                ]) { response, xml ->
-                            if (xml) {
-                                def authenticationError = getSushiErrorMessage(xml)
-                                if (authenticationError){
-                                    jsonErrors.add(authenticationError)
-                                    csr.jerror = JsonOutput.toJson(jsonErrors)
-                                }
-                                if (responseHasUsageData(xml, titleId)) {
-                                    def statsTitles = xml.depthFirst().findAll {
-                                        it.name() == 'ItemName'
-                                    }
-                                    if (statsTitles.size() > 1) {
-                                        log.warn('Found more than one item for the given Identifier')
-                                        log.warn('Titles delivered by API: ')
-                                        log.warn(statsTitles)
-                                    }
-                                    def itemPerformances = xml.depthFirst().findAll {
-                                        it.name() == 'ItemPerformance'
-                                    }
-                                    def usageMap = getPeriodUsageMap(itemPerformances)
-                                    def cal = new GregorianCalendar()
-
-                                    usageMap.each { key, countPerMetric ->
-                                        def fact = [:]
-                                        countPerMetric.each { metric, count ->
-                                            fact.from = timeStampFormat.parse(key)
-                                            fact.to =timeStampFormat.parse(getLastDayOfMonth(key))
-                                            cal.setTime(fact.to)
-                                            fact.reportingYear=cal.get(Calendar.YEAR)
-                                            fact.reportingMonth=cal.get(Calendar.MONTH)+1
-                                            fact.type = statsReport.toString()
-                                            fact.value = count
-                                            fact.uid = "${titleId}:${platform}:${customer}:${key}:${metric}:${statsReport.value}"
-                                            fact.metric = RefdataValue.getByValueAndCategory(metric,'FactMetric')
-                                            fact.title = title_inst
-                                            fact.supplier = supplier_inst
-                                            fact.inst = org_inst
-                                            fact.juspio = title_io_inst
-                                            if (factService.registerFact(fact)) {
-                                                ++factCount
-                                                ++newFactCount
-                                                DateTimeFormatter formatter = DateTimeFormatter.ofPattern('yyyy-MM-dd')
-                                                csr.haveUpTo = YearMonth.parse(key, formatter).toString()
-                                            }
-                                        }
-                                    }
-                                    log.debug("Title: ${title_inst.title}")
-                                    log.debug("ID: ${titleId}")
-
-                                }
-                            } else {
-                                def errorMessage = "No xml object returned, response status: ${response.statusLine}"
-                                log.error(errorMessage)
-                                jsonErrors.add(errorMessage)
-                                csr.jerror = JsonOutput.toJson(jsonErrors)
-                            }
+                        sushiClient.query()
+                        def xml = sushiClient.getResult()
+                        def authenticationError = getSushiErrorMessage(xml)
+                        if (authenticationError) {
+                            jsonErrors.add(authenticationError)
+                            csr.jerror = JsonOutput.toJson(jsonErrors)
                         }
-
+                        if (responseHasUsageData(xml, options.statsTitleIdentifier)) {
+                            writeUsageRecords(xml, options, csr)
+                        }
                     } catch (Exception e) {
                         log.error("Error fetching data")
                         log.error(e.message)
@@ -354,25 +274,7 @@ class StatsSyncService {
                             csr.jerror = jsonError
                         }
                     }
-                    csr.numFacts = (csr.numFacts) ? csr.numFacts + factCount : factCount
-                    try {
-                        csr.save(flush: true)
-                    } catch (Exception e) {
-                        log.error(e.message)
-                        jsonErrors.add(e.message)
-                        def jsonError = JsonOutput.toJson(jsonErrors)
-                        if (jsonError) {
-                            csr.jerror = jsonError
-                        }
-                        exceptionCount++
-                    }
                 }
-            }
-            // TODO remove?
-            // Exceptions are all catched, do we really want to save here when there were certain exceptions?
-            // For now save the csr which should contain an error message
-            if (csr != null) {
-                csr.save(flush: true)
             }
             cleanUpGorm()
             def elapsed = System.currentTimeMillis() - start_time;
@@ -381,12 +283,239 @@ class StatsSyncService {
         }
     }
 
-    private String getMostRecentClosedPeriod()
-    {
-        def c = new GregorianCalendar()
-        c.add(Calendar.MONTH,-2) // -2 TODO change back
-        // Remember months are zero based - hence the +1 in this line!
-        return "${c.get(Calendar.YEAR)}-${String.format('%02d',c.get(Calendar.MONTH)+1)}"
+    def writeUsageRecords(xml, options, csr) {
+        checkStatsTitleCount(xml)
+        def itemPerformances = xml.depthFirst().findAll {
+            it.name() == 'ItemPerformance'
+        }
+        if (itemPerformances.size()>0) {
+            itemPerformances.sort() {
+                it.Period.Begin.text()
+            }
+        }
+        // 3030 Exception, no usage data for fetched report
+        if (itemPerformances.empty) {
+            csr.availTo = new SimpleDateFormat('yyyy-MM').parse(options.mostRecentClosedPeriod)
+            csr.save(flush: true)
+            return
+        }
+        def usageRanges = getUsageRanges(itemPerformances, options, getNotProcessedMonths(xml))
+        def cal = new GregorianCalendar()
+        usageRanges.each {
+            def factCount = 0
+            def itemPerformancesForRange = getItemPerformancesForRange(itemPerformances, it)
+            csr.availFrom = new SimpleDateFormat('yyyy-MM').parse(it['begin'])
+            // should only happen on first sync if there is a range without usage before the first ItemPerformance, e.g. if
+            // we want to get usage for 2012ff from NatStat, but we cannot get usage this early
+            if (itemPerformancesForRange.empty) {
+                csr.availTo = new SimpleDateFormat('yyyy-MM-dd').parse(getDateForLastDayOfMonth(it['end']))
+                csr.save(flush: true)
+            } else {
+                def usageMap = getPeriodUsageMap(itemPerformancesForRange)
+                usageMap.each { key, countPerMetric ->
+                    def fact = [:]
+                    countPerMetric.each { metric, count ->
+                        fact.from = new SimpleDateFormat('yyyy-MM-dd').parse(key)
+                        fact.to = new SimpleDateFormat('yyyy-MM-dd').parse(getDateForLastDayOfMonth(key))
+                        cal.setTime(fact.to)
+                        fact.reportingYear = cal.get(Calendar.YEAR)
+                        fact.reportingMonth = cal.get(Calendar.MONTH) + 1
+                        fact.type = csr.factType.value
+                        fact.value = count
+                        fact.uid = "${options.statsTitleIdentifier}:${options.platform}:${options.customer}:${key}:${metric}:${csr.factType.value}"
+                        fact.metric = RefdataValue.getByValueAndCategory(metric, 'FactMetric')
+                        fact.title = options.title_inst
+                        fact.supplier = options.supplier_inst
+                        fact.inst = options.org_inst
+                        fact.juspio = options.title_io_inst
+                        if (factService.registerFact(fact)) {
+                            ++factCount
+                            ++newFactCount
+                            DateTimeFormatter formatter = DateTimeFormatter.ofPattern('yyyy-MM-dd')
+                        }
+                    }
+                }
+                // First csr -> update
+                if (csr.availTo == null){
+                    csr.availTo = new SimpleDateFormat('yyyy-MM').parse(it.end)
+                    csr.numFacts = factCount
+                    csr.save(flush: true)
+
+                } else {
+                    def newFromPeriod = getNextFromPeriod(csr).substring(0,7)
+                    if (newFromPeriod != it.begin) { // gap for new range
+                        log.warn("usage data gap found before ${it.begin}")
+                    }
+                    csr = new StatsTripleCursor()
+                    csr.availFrom = new SimpleDateFormat('yyyy-MM').parse(it.begin) // update
+                    csr.availTo = new SimpleDateFormat('yyyy-MM').parse(getDateForLastDayOfMonth(it['end'])) // update to last month for that range
+                    csr.customerId = options.customer
+                    csr.numFacts = factCount
+                    csr.titleId = options.statsTitleIdentifier
+                    csr.supplierId = options.platform
+                    csr.factType = options.factType
+                    csr.save(flush: true)
+                }
+            }
+        }
+    }
+
+    def getItemPerformancesForRange(itemPerformances, range) {
+        itemPerformances.findAll {
+            it.Period.Begin.text().substring(0,7) >= range["begin"] &&
+                it.Period.End.text().substring(0,7) <= range["end"]
+        }
+    }
+
+    def getRangeBeforeFirstItemPerformanceElement(itemPerformances, options, notProcessedMonths) {
+        def rangeMap = [:]
+        def firstItemPerformanceBeginPeriod = itemPerformances.first().Period.Begin.text().substring(0,7)
+        def lastItemPerformanceBeginPeriod = itemPerformances.last().Period.Begin.text().substring(0,7)
+        def firstProcessedMonth = null
+        def monthsWithoutNatStatTasksBeforeFirstItemPerformancePeriod = notProcessedMonths.findAll {
+            firstItemPerformanceBeginPeriod > it
+        }
+        if (monthsWithoutNatStatTasksBeforeFirstItemPerformancePeriod) {
+            DateTimeFormatter formatter = DateTimeFormatter.ofPattern('yyyy-MM')
+            YearMonth localDate = YearMonth.parse(monthsWithoutNatStatTasksBeforeFirstItemPerformancePeriod.last(),
+                formatter)
+            firstProcessedMonth = localDate.plusMonths(1).toString()
+        }
+        if (firstProcessedMonth && (firstProcessedMonth < firstItemPerformanceBeginPeriod)) {
+            rangeMap['begin'] = firstProcessedMonth
+            DateTimeFormatter formatter = DateTimeFormatter.ofPattern('yyyy-MM')
+            YearMonth localDate = YearMonth.parse(firstItemPerformanceBeginPeriod,formatter)
+            rangeMap['end'] =  localDate.minusMonths(1).toString()
+        }
+        return rangeMap
+    }
+
+    def getUsageRanges(ArrayList itemPerformances, options, notProcessedMonths) {
+        log.debug('Get Usage ranges for API call from/to Period')
+        def ranges = []
+        // Add begin end period for zero usage before first ItemPerformance and filter out months not available before that.
+        // At the moment we begin SYNC_STATS_FROM. We would have to extend the NatStat API to improve that (i.e.
+        // get csr.availfrom for the first period via API call
+        def rangeMap = [:]
+        def rangeBeforeFirstItemPerformanceElement = getRangeBeforeFirstItemPerformanceElement(itemPerformances,
+            options, notProcessedMonths)
+
+        if (rangeBeforeFirstItemPerformanceElement.size() != 0) {
+            // There is zero usage for one or more months before the first ItemPerformance (range between last not processed month
+            // and first ItemPerformance Period), add those months to range
+            rangeMap['begin'] = rangeBeforeFirstItemPerformanceElement['begin']
+            log.debug('Found months between not processed months (SUSHI 3031) and first ItemPerformance Period')
+        } else {
+            // use next from period which was used to query the SUSHI service
+            rangeMap['begin'] = options.from
+            // Case when we have only not processed (3031) Months before first ItemPerformance
+            // we have to set rangeMap['begin'] to +1 month after (last not processed month before first ItemPerformance month)
+            def lastNotProcessedMonthBeforeItemPerformances = notProcessedMonths.findAll() {
+                it < itemPerformances.first().Period.Begin.text().substring(0,7)
+            }.sort()
+            if (lastNotProcessedMonthBeforeItemPerformances.size() != 0) {
+                DateTimeFormatter formatter = DateTimeFormatter.ofPattern('yyyy-MM')
+                YearMonth localDate = YearMonth.parse(lastNotProcessedMonthBeforeItemPerformances.last(), formatter)
+                rangeMap['begin'] = localDate.plusMonths(1).toString()
+            }
+
+            if (itemPerformances.first().Period.Begin.text().substring(0,7) != options.from.substring(0,7)){
+                log.debug('First ItemPerformance does not equal from period for SUSHI call')
+            }
+        }
+
+        // remove not processed months which we already have used to check for gaps to be able to use the list to find newer gaps
+        // also handle the case where we have only 3031 Months before
+        notProcessedMonths.removeAll {
+            it < rangeMap['begin']
+        }
+        def itemPerformanceRangeList = [] //temp list
+        itemPerformances.each { performance ->
+            def performanceMonth = performance.Period.Begin.text().substring(0,7)
+            // we have a gap if there is a month in our list < the actual processed ItemPerformance month
+            def gap = notProcessedMonths.find { month ->
+                 performanceMonth > month
+            }
+
+            // if we have a gap, close range and remove all not processed months belonging together and also within that gap
+            if (gap){
+                if (itemPerformanceRangeList.size() == 0){
+                    log.error('Gap but no Itemperformance Elements to calculate the end of range')
+                    // throw exception?
+                }
+                rangeMap['end'] = itemPerformanceRangeList.last().Period.Begin.text().substring(0,7)
+                ranges.add(rangeMap)
+                notProcessedMonths.removeAll {
+                    it < performanceMonth
+                }
+                rangeMap = [:]
+                rangeMap['begin'] = performanceMonth
+                itemPerformanceRangeList = []
+                itemPerformanceRangeList.add(performance)
+
+            } else {
+                itemPerformanceRangeList.add(performance)
+            }
+        }
+        rangeMap['end'] = itemPerformanceRangeList.last().Period.Begin.text().substring(0,7)
+        ranges.add(rangeMap)
+
+        return ranges
+    }
+
+    /**
+     * Not processed months when getting a 3031 Exception
+     * @param xml
+     * @return
+     */
+    def getNotProcessedMonths(xml) {
+        if (xml.Exception.isEmpty() == false && xml.Exception.Number == '3031') {
+            def exceptionData = xml.Exception.Data
+            def matcher = exceptionData =~ /\d{4}-\d{2}/
+            def list = []
+            matcher.each {
+                list.add(it)
+            }
+            return list
+        }
+        return []
+    }
+
+
+    def checkStatsTitleCount(xml) {
+        def statsTitles = xml.depthFirst().findAll {
+            it.name() == 'ItemName'
+        }
+        if (statsTitles.size() > 1) {
+            log.warn('Found more than one item for the given Identifier')
+            log.warn('Titles delivered by API: ')
+            log.warn(statsTitles)
+        }
+    }
+
+    def getNextFromPeriod(csr) {
+        def acceptedFormat = "yyyy-MM-dd"
+        def fromPeriodForAPICall
+        // Latest stored month + 1
+        if (csr.availTo) {
+            use(TimeCategory) {
+                fromPeriodForAPICall = csr.availTo + 1.month
+            }
+        } else {
+            if (!csr.availFrom){
+                return SYNC_STATS_FROM
+            }
+            fromPeriodForAPICall = csr.availFrom
+        }
+        return fromPeriodForAPICall.format(acceptedFormat)
+    }
+
+    private String getMostRecentClosedPeriod() {
+        Calendar cal = Calendar.getInstance();
+        cal.setTime(new Date());
+        cal.add(Calendar.MONTH, -2)
+        cal.set(Calendar.DAY_OF_MONTH, cal.getActualMaximum(Calendar.DAY_OF_MONTH))
+        return cal.format('yyyy-MM-dd')
     }
 
     private isAllowedMetric(metric) {
@@ -425,7 +554,7 @@ class StatsSyncService {
         return map
     }
 
-    private String getLastDayOfMonth(yearMonthString) {
+    private String getDateForLastDayOfMonth(yearMonthString) {
         def sdf = new SimpleDateFormat("yyyy-MM")
         def cal = new GregorianCalendar()
         cal.setTime(sdf.parse(yearMonthString))
@@ -433,7 +562,7 @@ class StatsSyncService {
         return "${cal.get(Calendar.YEAR)}-${String.format('%02d',cal.get(Calendar.MONTH)+1)}-${cal.get(Calendar.DAY_OF_MONTH)}"
     }
 
-    private getSushiErrorMessage(xml){
+    private getSushiErrorMessage(xml) {
         if (xml.Exception.isEmpty() == false) {
             def errorNumber = xml.Exception.Number
             def sushiErrorList = ['2000', '2020', '3000', '3062']
@@ -446,10 +575,14 @@ class StatsSyncService {
 
     private Boolean responseHasUsageData(xml, titleId) {
         // TODO maybe better check for usage first
-        // What if we get a 3030 Exception? We return false here and do not store facts for the queried period.
+        // What if we get a 3030 Exception? We return true here (report is processed in NatStat, usage data = 0)
+
         // Do we need to handle the 3031 Exceptions and kind of flag periods which are mentioned in the XML data element?
         // Or store 0 values for 3030 Exceptions, which allows us to mark missing/errorneous months, but would increase
         // the number of facts significantly
+        if (xml.Exception.isEmpty() == false && xml.Exception.Number == '3030'){
+            return true
+        }
         if (xml.Exception.isEmpty() == false && xml.Exception.Number != '3031') {
             log.debug('SUSHI Exception Number ' + xml.Exception.Number + ' : ' + xml.Exception.Message)
             return false
@@ -463,6 +596,7 @@ class StatsSyncService {
             return true
         }
     }
+
 
     def cleanUpGorm() {
         log.debug("Clean up GORM")